Movement Scripts Added
I was able to get movement enabled on one of the paddles.

However, the paddle only moves in the x-axis.

I believe this is in part, to the way the CharacterController2D script was made.

The assumption of that script is that the moving object is running in the x-axis, like a platforming game.

I have to look deeper in the code to see how to fix it.

I may even consult the documentation and build a script from scratch.